refactor(code-topology): move force_3d viz to template and split viz_3d into model / layout / render
* Extracts the force_3d dashboard viz HTML body into
  templates/force_3d.hbs (loaded via include_str!).
* Turns viz_3d.rs into a directory module:
    - viz_3d/model.rs    typed config and scene graph
    - viz_3d/layout.rs   pure force-directed simulation
    - viz_3d/render.rs   Scene3D construction plus Three.js viewer render
    - viz_3d/templates/scene.hbs   Three.js HTML template
    - viz_3d/mod.rs      public ForceDirectedProjector and Projector impl
  Public API surface (ForceDirectedProjector, ForceDirectedConfig,
  Scene3D and friends, register) is unchanged.

Behavior preserved: baseline diff for topology-3d.html and the other
viz outputs is only the trailing newline that include_str! adds.

Refs #108.
